+++ Cargo.toml
@@ -5,12 +5,12 @@ members = [
     "fe2o3-amqp",
     "fe2o3-amqp-ext",
     "fe2o3-amqp-types", 
-    "fe2o3-amqp-ws",
     "fe2o3-amqp-management",
     "fe2o3-amqp-cbs",
 ]
 
 exclude = [
+    "fe2o3-amqp-ws", # TODO: move back to members after wasm32 experiment
     "examples/alternative_tls_connection",
     "examples/batchable_send",
     "examples/dispose_multiple",
