+++ fe2o3-amqp-ws/src/lib.rs
@@ -66,8 +66,6 @@
 //! }
 //! ```
 
-const SEC_WEBSOCKET_PROTOCOL: &str = "Sec-WebSocket-Protocol";
-
 use std::{
     io::{self, Cursor, Read},
     task::Poll,
@@ -77,22 +75,25 @@ use futures_util::{ready, Sink, Stream};
 use pin_project_lite::pin_project;
 use tokio::{
     io::{AsyncRead, AsyncWrite},
-    net::TcpStream,
-};
-use tokio_tungstenite::{
-    client_async, client_async_with_config, connect_async, connect_async_with_config,
-    MaybeTlsStream,
-};
-use tungstenite::{
-    client::IntoClientRequest,
-    handshake::client::{Request, Response},
-    http::HeaderValue,
-    protocol::WebSocketConfig,
-    Message,
 };
 
 mod error;
 pub use error::Error;
+
+#[macro_use]
+mod macros;
+
+cfg_not_wasm32! {
+    mod native;
+}
+
+cfg_wasm32! {
+    mod wasm;
+}
+
+/// This a wrapper around `tungstenite::Message`
+#[derive(Debug, Eq, PartialEq, Clone)]
+pub struct WsMessage(pub tungstenite::Message);
 
 pin_project! {
     /// A wrapper over [`tokio_tungstenite::WebSoccketStream`] that implements
@@ -139,185 +140,8 @@ pin_project! {
     #[derive(Debug)]
     pub struct WebSocketStream<S> {
         #[pin]
-        inner: tokio_tungstenite::WebSocketStream<S>,
+        inner: S,
         current_binary: Option<std::io::Cursor<Vec<u8>>>,
-    }
-}
-
-impl<S> From<tokio_tungstenite::WebSocketStream<S>> for WebSocketStream<S> {
-    fn from(inner: tokio_tungstenite::WebSocketStream<S>) -> Self {
-        Self {
-            inner,
-            current_binary: None,
-        }
-    }
-}
-
-impl WebSocketStream<MaybeTlsStream<TcpStream>> {
-    /// Calls [`tokio_tungstenite::connect_async`] internally with `"Sec-WebSocket-Protocol"` HTTP
-    /// header of the `req` set to `"amqp"`
-    pub async fn connect(req: impl IntoClientRequest) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) = connect_async(request).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-
-    /// Calls [`tokio_tungstenite::connect_async_with_config`] internally with
-    /// `"Sec-WebSocket-Protocol"` HTTP header of the `req` set to `"amqp"`
-    pub async fn connect_with_config(
-        req: impl IntoClientRequest,
-        config: Option<WebSocketConfig>,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) = connect_async_with_config(request, config).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-}
-
-impl<S> WebSocketStream<S>
-where
-    S: AsyncRead + AsyncWrite + Unpin,
-{
-    /// Calls [`tokio_tungstenite::client_async`] internally with `"Sec-WebSocket-Protocol"` HTTP
-    /// header of the `req` set to `"amqp"`
-    pub async fn connect_with_stream(
-        req: impl IntoClientRequest,
-        stream: S,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) = client_async(request, stream).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-
-    /// Calls [`tokio_tungstenite::client_async_with_config`] internally with
-    /// `"Sec-WebSocket-Protocol"` HTTP header of the `req` set to `"amqp"`
-    pub async fn connect_with_stream_and_config(
-        req: impl IntoClientRequest,
-        stream: S,
-        config: Option<WebSocketConfig>,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) = client_async_with_config(request, stream, config).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-}
-
-#[cfg_attr(
-    docsrs,
-    doc(cfg(any(
-        feature = "native-tls",
-        feature = "native-tls-vendored",
-        feature = "rustls-tls-native-roots",
-        feature = "rustls-tls-webpki-roots"
-    )))
-)]
-#[cfg(any(
-    feature = "native-tls",
-    feature = "native-tls-vendored",
-    feature = "rustls-tls-native-roots",
-    feature = "rustls-tls-webpki-roots"
-))]
-impl<S> WebSocketStream<MaybeTlsStream<S>>
-where
-    S: AsyncRead + AsyncWrite + Send + Sync + Unpin + 'static,
-{
-    /// Calls [`tokio_tungstenite::client_async_tls`] internally with `"Sec-WebSocket-Protocol"` HTTP
-    /// header of the `req` set to `"amqp"`
-    pub async fn connect_tls_with_stream(
-        req: impl IntoClientRequest,
-        stream: S,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) =
-            tokio_tungstenite::client_async_tls(request, stream).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-
-    /// Calls [`tokio_tungstenite::client_async_tls_with_config`] internally with
-    /// `"Sec-WebSocket-Protocol"` HTTP header of the `req` set to `"amqp"`
-    pub async fn connect_tls_with_stream_and_config(
-        req: impl IntoClientRequest,
-        stream: S,
-        config: Option<WebSocketConfig>,
-        connector: Option<tokio_tungstenite::Connector>,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) =
-            tokio_tungstenite::client_async_tls_with_config(request, stream, config, connector)
-                .await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
-    }
-}
-
-#[cfg_attr(
-    docsrs,
-    doc(cfg(any(
-        feature = "native-tls",
-        feature = "native-tls-vendored",
-        feature = "rustls-tls-native-roots",
-        feature = "rustls-tls-webpki-roots"
-    )))
-)]
-#[cfg(any(
-    feature = "native-tls",
-    feature = "native-tls-vendored",
-    feature = "rustls-tls-native-roots",
-    feature = "rustls-tls-webpki-roots"
-))]
-impl WebSocketStream<MaybeTlsStream<TcpStream>> {
-    /// Calls [`tokio_tungstenite::connect_async_tls_with_config`] internally with
-    /// `"Sec-WebSocket-Protocol"` HTTP header of the `req` set to `"amqp"`
-    pub async fn connect_tls_with_config(
-        req: impl IntoClientRequest,
-        config: Option<WebSocketConfig>,
-        connector: Option<tokio_tungstenite::Connector>,
-    ) -> Result<(Self, Response), Error> {
-        let request = map_amqp_websocket_request(req)?;
-        let (mut ws_stream, response) =
-            tokio_tungstenite::connect_async_tls_with_config(request, config, connector).await?;
-        match verify_response(response) {
-            Ok(response) => Ok((Self::from(ws_stream), response)),
-            Err(error) => {
-                ws_stream.close(None).await?;
-                Err(error)
-            }
-        }
     }
 }
 
@@ -328,7 +152,7 @@ impl WebSocketStream<MaybeTlsStream<TcpStream>> {
 // - `ws_stream_tungstenite`
 impl<S> AsyncRead for WebSocketStream<S>
 where
-    S: AsyncRead + AsyncWrite + Unpin,
+    S: Stream<Item = Result<WsMessage, tungstenite::Error>>,
 {
     fn poll_read(
         self: std::pin::Pin<&mut Self>,
@@ -353,24 +177,24 @@ where
                 None => return Poll::Ready(Ok(())), // EOF
             };
 
-            match msg {
-                Message::Text(_) => {
+            match msg.0 {
+                tungstenite::Message::Text(_) => {
                     return Poll::Ready(Err(io::Error::new(
                         io::ErrorKind::InvalidData,
                         "Text messsage is not supported",
                     )))
                 }
-                Message::Binary(vec) => *this.current_binary = Some(Cursor::new(vec)),
+                tungstenite::Message::Binary(vec) => *this.current_binary = Some(Cursor::new(vec)),
 
                 // This is already handled by tungstenite
-                Message::Ping(_) => {}
-                Message::Pong(_) => {}
+                tungstenite::Message::Ping(_) => {}
+                tungstenite::Message::Pong(_) => {}
 
                 // Let tungstenite perform close handshake
-                Message::Close(_) => {}
+                tungstenite::Message::Close(_) => {}
 
                 // Raw frame. Note, that you’re not going to get this value while reading the message.
-                Message::Frame(_) => unreachable!(),
+                tungstenite::Message::Frame(_) => unreachable!(),
             }
         };
 
@@ -387,7 +211,7 @@ where
 
 impl<S> AsyncWrite for WebSocketStream<S>
 where
-    S: AsyncRead + AsyncWrite + Unpin,
+    S: Sink<WsMessage, Error = tungstenite::Error>,
 {
     fn poll_write(
         self: std::pin::Pin<&mut Self>,
@@ -397,7 +221,8 @@ where
         let mut this = self.project();
         ready!(this.inner.as_mut().poll_ready(cx)).map_err(map_tungstenite_error)?;
         let n = buf.len();
-        let item = Message::binary(buf);
+        let item = tungstenite::Message::binary(buf);
+        let item = WsMessage(item);
         match this.inner.start_send(item) {
             Ok(_) => Poll::Ready(Ok(n)),
             Err(error) => Poll::Ready(Err(map_tungstenite_error(error))),
@@ -429,41 +254,5 @@ fn map_tungstenite_error(error: tungstenite::Error) -> io::Error {
         tungstenite::Error::Io(err) => err,
         tungstenite::Error::Capacity(err) => io::Error::new(io::ErrorKind::InvalidData, err),
         _ => io::Error::new(io::ErrorKind::Other, error),
-    }
-}
-
-fn map_amqp_websocket_request(req: impl IntoClientRequest) -> Result<Request, tungstenite::Error> {
-    let mut request = req.into_client_request()?;
-
-    // Sec-WebSocket-Protocol HTTP header
-    //
-    // Identifies the WebSocket subprotocol. For this AMQP WebSocket binding, the value MUST be
-    // set to the US- ASCII text string “amqp” which refers to the 1.0 version of the AMQP 1.0
-    // or greater, with version negotiation as defined by AMQP 1.0.
-    request
-        .headers_mut()
-        .insert(SEC_WEBSOCKET_PROTOCOL, HeaderValue::from_static("amqp"));
-
-    Ok(request)
-}
-
-fn verify_response(response: Response) -> Result<Response, Error> {
-    use http::StatusCode;
-
-    // If the Client does not receive a response with HTTP status code 101 and an HTTP
-    // Sec-WebSocket-Protocol equal to the US-ASCII text string “amqp” then the Client MUST close
-    // the socket connection
-    if response.status() != StatusCode::SWITCHING_PROTOCOLS {
-        return Err(Error::StatucCodeIsNotSwitchingProtocols);
-    }
-
-    match response
-        .headers()
-        .get(SEC_WEBSOCKET_PROTOCOL)
-        .map(|val| val.to_str())
-        .ok_or(Error::MissingSecWebSocketProtocol)??
-    {
-        "amqp" => Ok(response),
-        _ => Err(Error::SecWebSocketProtocolIsNotAmqp),
     }
 }
