+++ fe2o3-amqp-ws/src/macros.rs
@@ -1,0 +1,17 @@
+macro_rules! cfg_wasm32 {
+    ($($item:item)*) => {
+        $(
+            #[cfg(target_arch = "wasm32")]
+            $item
+        )*
+    };
+}
+
+macro_rules! cfg_not_wasm32 {
+    ($($item:item)*) => {
+        $(
+            #[cfg(not(target_arch = "wasm32"))]
+            $item
+        )*
+    }
+}
