+++ fe2o3-amqp-ws/src/message.rs
@@ -1,0 +1,2 @@
+//! WebSocket message
+
