+++ fe2o3-amqp-ws/src/wasm.rs
@@ -1,0 +1,3 @@
+use web_sys::{WebSocket, MessageEvent};
+
+struct WasmWebSocketStream(WebSocket);
